+++ src/FluentSerializer.Json/Converting/Converters/DateTimeOffsetByFormatConverter.cs
@@ -1,3 +1,6 @@
+#if NETSTANDARD2_0
+using FluentSerializer.Core.Dirty.BackwardsCompatibility.NetFramework;
+#endif
 using Ardalis.GuardClauses;
 
 using FluentSerializer.Json.Converting.Converters.Base;
@@ -23,12 +26,12 @@ public class DateTimeOffsetByFormatConverter : SimpleTypeConverter<DateTimeOffse
 	public DateTimeOffsetByFormatConverter(in string format, in CultureInfo cultureInfo, in DateTimeStyles dateTimeStyle)
 	{
 		Guard.Against.NullOrWhiteSpace(format
-#if NETSTANDARD2_1
+#if NETSTANDARD
 			, nameof(format)
 #endif
 		);
 		Guard.Against.Null(cultureInfo
-#if NETSTANDARD2_1
+#if NETSTANDARD
 			, nameof(cultureInfo)
 #endif
 		);