+++ src/FluentSerializer.Json/DataNodes/Nodes/JsonProperty.Writing.cs
@@ -22,7 +22,7 @@ public readonly partial struct JsonProperty
 	public ITextWriter AppendTo(ref ITextWriter stringBuilder, in bool format = true, in int indent = 0, in bool writeNull = true)
 	{
 		Guard.Against.NullOrWhiteSpace(Name, message: "The property was is an illegal state, it contains no Name"
-#if NETSTANDARD2_1
+#if NETSTANDARD
 			, parameterName: nameof(Name)
 #endif
 		);