+++ src/FluentSerializer.Xml/Converting/Converters/DateTimeByFormatConverter.cs
@@ -22,12 +22,12 @@ public class DateTimeByFormatConverter : SimpleTypeConverter<DateTime>
 	public DateTimeByFormatConverter(in string format, in CultureInfo cultureInfo, in DateTimeStyles dateTimeStyle)
 	{
 		Guard.Against.NullOrWhiteSpace(format
-#if NETSTANDARD2_1
+#if NETSTANDARD
 			, nameof(format)
 #endif
 		);
 		Guard.Against.Null(cultureInfo
-#if NETSTANDARD2_1
+#if NETSTANDARD
 			, nameof(cultureInfo)
 #endif
 		);