+++ src/FluentSerializer.Xml/Converting/Converters/DateTimeOffsetByFormatConverter.cs
@@ -22,12 +22,12 @@ public class DateTimeOffsetByFormatConverter : SimpleTypeConverter<DateTimeOffse
 	public DateTimeOffsetByFormatConverter(in string format, in CultureInfo cultureInfo, in DateTimeStyles dateTimeStyle)
 	{
 		Guard.Against.NullOrWhiteSpace(format
-#if NETSTANDARD2_1
+#if NETSTANDARD
 			, nameof(format)
 #endif
 		);
 		Guard.Against.Null(cultureInfo
-#if NETSTANDARD2_1
+#if NETSTANDARD
 			, nameof(cultureInfo)
 #endif
 		);