+++ src/FluentSerializer.Xml.DependencyInjection.NetCoreDefault/Changelog.md
@@ -2,9 +2,9 @@
 
 ## @next
 
--
+- [#8](https://github.com/Marvin-Brouwer/FluentSerializer/issues/8) Added `netframework` backwards compatibility
 
-## 2.0.2
+## v2.0.2
 
 - [#146](https://github.com/Marvin-Brouwer/FluentSerializer/issues/146) Added `net7` support